% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipool-package.R
\docType{package}
\name{epipool-package}
\alias{epipool-package}
\alias{epipool}
\title{The 'epipool' package.}
\description{
This package provides Bayesian models estimating a prevalence over time from pooled test results.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
