% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolprev1.R
\name{poolprev1}
\alias{poolprev1}
\title{Estimate prevalence from pooled test results assuming a fixed prevalence parameter for each time point}
\usage{
poolprev1(
  data,
  method = "GP",
  spec = 0.995,
  prior = list(lengthscale = c(0, 1), sd_GP = c(0, 1), logit_prev = c(-4, 2)),
  ...
)
}
\arguments{
\item{data}{Datasets with the following variables:

\itemize{
\item time. Time (should be numeric)
\item pool.size. Pool size (number of samples by pool)
\item n.pools. Number of pools
\item n.pos.pools. Number of positive pools.
}}

\item{method}{Method used to analyse data (timepoint or GP)}

\item{spec}{Test specificity}

\item{prior}{list of values for hyperprior parameters.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Estimate prevalence from pooled test results assuming a fixed prevalence parameter for each time point
}
\examples{
#load data
data(epipool_data1)

#run model with method assuming a prevalence parameter every week
prev_timepoint = poolprev1(epipool_data1[[1]],method="timepoint") #takes a few seconds
#run model with GP
prev_GP = poolprev1(epipool_data1[[1]],method="GP",control=list(adapt_delta=0.99,max_treedepth=12)) #can take a few minutes

#plot
library(ggplot2)
left_join(epipool_data1[[1]],
rbind(prev_timepoint \%>\% dplyr::mutate(method="timepoint"),
      prev_GP \%>\% dplyr::mutate(method="GP")),
      by="time") \%>\%
ggplot(aes(x=time))+
geom_ribbon(aes(ymin=lwr,ymax=upr),fill="black",alpha=0.1)+
geom_line(aes(y=mean),col="black",size=0.8)+
geom_point(aes(y=mean),col="black",size=2)+
geom_point(aes(y=prev),col="red",size=2) +
facet_grid(.~method) +
theme_bw()
}
